+++ designer/restmodel/src/main/scala/pl/touk/nussknacker/restmodel/BaseEndpointDefinitions.scala
@@ -15,7 +15,7 @@ trait BaseEndpointDefinitions {
   val baseNuApiEndpoint: PublicEndpoint[Unit, Unit, Unit, Any] = endpoint.in("api")
 
   implicit def toSecuredEndpoint[INPUT, BUSINESS_ERROR, OUTPUT, R](
-      endpoint: Endpoint[Unit, INPUT, BUSINESS_ERROR, OUTPUT, R]
+      endpoint: PublicEndpoint[INPUT, BUSINESS_ERROR, OUTPUT, R]
   ): ToSecure[INPUT, BUSINESS_ERROR, OUTPUT, R] =
     new ToSecure(endpoint)